Fixed 1.20.2 crashes caused by staticdependent.json
Having modify attribute for joltik's base health instead of a plain un-conditioned attribute put the player in a death loop upon choosing joltik that could only be fixed by removing the datapack/joltik origin so that the player's origin would be reset, and re-adding it after the player is no longer joltik. This power caused no issues on 1.20/1.20.1.