just making sure any changes that got made are in.
i dont remember if there were any. this power is fragile. it was a joint effort by myself and rightupmyallie./idk.polar after months of my struggling alone beforehand. we do not fuck with it. if it works, it works. if it doesn't, i'll light a candle and grab the rubber duck.